requirements.txt
google-auth==2.20.0
google-auth-oauthlib==0.5.3
google-auth-httplib2==0.1.0
redis==4.3.4
apscheduler==3.9.1

Flask==2.2.5
Flask-Dance==7.1.0
Flask-Assets==2.0
Flask-SQLAlchemy==3.0.3
oauthlib==3.2.2
requests-oauthlib==1.3.1
python-dateutil==2.9.0.post0
gunicorn==21.2.0
google-api-python-client==2.70.0
psycopg2-binary>=2.9
